% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de2cca.R
\name{de2cca}
\alias{de2cca}
\title{Cannonical correspondence analysis (CCA) of gene expression}
\usage{
de2cca(info, counts, fast = TRUE, formula, palette = NULL, ...)
}
\arguments{
\item{info}{An experimental design matrix}

\item{counts}{A count matrix}

\item{fast}{If the dataset has >10k genes, this indicates that the expression matrix should be
subsampled down to 10k. Datasets with >10k genes can take >10 minutes to run}

\item{formula}{A character string that can be coerced to a formula. Terms must be separated
by a whitespace}

\item{palette}{A R palette that is used to color the dendrogram}
}
\value{
a vegan::cca model object
}
\description{
\code{de2cca} A combined multivariate analysis of gene expression that invokes both
canonical correspondence analysis of the expression matrix and a hierarchical clustering
of the distance matrix.
}
\examples{
library(PhyGenomicsData.PV2016)
data(info12); data(counts12)
info<-info12
counts<-counts12
test<-de2cca(info=info, counts=counts[1:1000,], formula="Treatment + MDWP + order")
}

