% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distGO.R
\name{distGO}
\alias{distGO}
\title{Compare semantic similarity of GO.}
\usage{
distGO(go.list, n = 10, ont = "BP", organism = "arabidopsis",
  measure = "Wang", combine = "rcmax", verbose = TRUE,
  calc.distance = TRUE)
}
\arguments{
\item{go.list}{A named list, where each element stores a character vector of GO
ids. GO terms should be sorted so that the most important are first.}

\item{n}{The number of GO terms to evaluate from the begining of the vector.}

\item{ont}{The ontology type: one of "MF", "BP", and "CC" subontologies to be
passed to mgoSim}

\item{organism}{Organism with GO database, to be passed to mgoSim}

\item{measure}{Measure of similarity to be passed to mgoSim}

\item{combine}{Measure of combining similarity to be passed to mgoSim}

\item{verbose}{Should updates be printed to the console?}

\item{calc.distance}{Should a distance matrix (default) or a similarity
matrix be returned.}

\item{...}{Not currently in use}
}
\value{
a matrix of (dis)similiarities
}
\description{
\code{distGO} A wrapper for GOSemSim::mgoSim, where similarity among
GO terms stored in a list are returned as a matrix
}
\details{
More here soon.
}
\examples{
\dontrun{
more here soon
}
}

