% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBinarySig.R
\name{makeBinarySig}
\alias{makeBinarySig}
\title{Convert p-values to binary significance calls}
\usage{
makeBinarySig(x, alpha = 0.05, what = "q.value", na.include = TRUE,
  verbose = TRUE)
}
\arguments{
\item{x}{A dataframe or matrix containing the significance (p/q/fdr p) values to convert}

\item{alpha}{The threshold that defines what is and is not significance}

\item{what}{Character vector defining a string to look for in the column names. For example, "q.value"}

\item{na.include}{Logical, should NAs be classified as not significant? If false, all rows with any NAs are dropped.}

\item{verbose}{Logical, should the number of significant values be printed}
}
\value{
a dataframe with all values in the original matrix with column names containing what, where
the values have been transformed to binary
}
\description{
\code{makeBinarySig} Convert some sort of significance distribution to binary signifcance calls.
}
\details{
Given a matrix with p/q values, this function greps for the string in "what" and classifies each column as
either significant (1) or not (0). Esspecially useful for a dataset that has many p-value or q-value columns
}
\examples{
data(kidney)
counts<-kidney$counts
counts<-counts[sample(1:nrow(counts),1000),]
info<-data.frame(rep=kidney$replic, treatment=kidney$treatment)
stats<-pipeLIMMA(counts=counts, info=info, formula = " ~ treatment", block=NULL)
sig<-makeBinarySig(x= stats$stats, what="p.value")
}

