% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means4heatmap.R
\name{means4heatmap}
\alias{means4heatmap}
\title{Get group means}
\usage{
means4heatmap(v, grps = info$Treatment)
}
\arguments{
\item{v}{A dataframe or matrix containing the (normalized) counts}

\item{grps}{The experimental factor to use to calculate groups.}
}
\value{
a matrix with means / gene.
}
\description{
\code{means4heatmap} Take normalized (or not)
counts data and produce a matrix that
can be turned into a heatmap with mean
values for each level.
}

