% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesByConnectivity.R
\name{modulesByConnectivity}
\alias{modulesByConnectivity}
\title{Find modules that have high connectivity}
\usage{
modulesByConnectivity(datExpr, net, mean.kme.threshold = seq(0, 1, 0.1))
}
\arguments{
\item{datExpr}{The expression dataset, transposed so that genes are columns
and individuals are rows.}

\item{net}{WGCNA generated network. Usually from WGCNA::blockwiseModules}

\item{mean.kme.threshold}{Numeric, the thresholds to test.
Can be of length >1.}
}
\value{
A vector of modules for each kme threshold specified.
}
\description{
\code{modulesByConnectivity} Determine the strongest modules in a network
using the mean absolute value of signedKME.
}
\details{
More here soon.
}
\examples{
\dontrun{
library(WGCNA)
library(igraph)
data(kidney) #' from simseq
counts<-kidney$counts
counts<-counts[sample(1:nrow(counts),1000),]
info<-with(kidney,
           data.frame(id = paste(replic, treatment, sep = "_"),
                      rep=replic,
                      Treatment=ifelse(treatment == "Tumor","tumor","cntr"),
                      stringsAsFactors=F))
colnames(counts)<-info$id
stats <- pipeLIMMA(counts = counts,
                   info = info,
                   block = NULL,
                   formula = "~ Treatment")

datExpr.1=t(stats$voom$E)
pow=6
net.1 = blockwiseModules(datExpr.1, power = pow,
                         maxBlockSize = 10000, deepSplit = 2,
                         minModuleSize = 10,
                         saveTOMs = FALSE,
                         verbose = F)
modulesByConnectivity(net = net.1, datExpr = datExpr.1)
}
}

