% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeLIMMA.R
\name{pipeLIMMA}
\alias{pipeLIMMA}
\title{A pipeline for LIMMA.}
\usage{
pipeLIMMA(counts, info, formula, block = NULL, design = NA,
  use.qualityWeights = TRUE, geneIDs = NA, getTopTable = FALSE,
  getEbayes = TRUE, simplify = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{counts}{A count matrix}

\item{info}{An experimental design matrix}

\item{formula}{A character string that can be coerced to a formula}

\item{block}{A string that represents an individual that was repeatedly measured, if NULL, runs the analysis without a blocking / duplicate correlation factor}

\item{use.qualityWeights}{Logical, run voom with quality weights or not?}

\item{geneIDs}{The names of genes. If NA, use row names from counts matrix}

\item{getTopTable}{Logical, return toptable statistics?}

\item{getEbayes}{Logical, return ebayes statistics?}

\item{simplify}{Logical, return a element with the F-statistics from the main model?}

\item{verbose}{Logical, return progress updates?}
}
\value{
a list with 4 or 5 (if simple=TRUE)
\itemize{
 \item{"stats"}{: the statsistics generated from ebayes, toptable, or both}
 \item{"voom"}{: the voom normalized counts data}
 \item{"lmfit"}{: the fitted model}
 \item{"countsSize"}{: the normalization factors for each library}
 \item{"simpleStats"}{: if simplify=TRUE, a dataset with the F statistics}
}
}
\description{
\code{pipeLIMMA} Run a pipeline of LIMMA functions for differential gene expression.
}
\details{
This function runs the following pipeline:
\itemize{
 \item{1. }{calculate normalization factors via edgeR::calcNormFactors}
 \item{2. }{Run limma::voom transformation}
 \item{3. }{Run limma:lmFit linear modeling}
 \item{4. }{Run limma::ebayes statistical modeling}
 \item{5. }{Ouput statistics and other data}
}
}
\examples{
library(SimSeq)
library(limmaDE2)
data(kidney)
counts<-kidney$counts
counts<-counts[sample(1:nrow(counts),1000),]
info<-data.frame(rep=kidney$replic, treatment=kidney$treatment)
stats<-pipeLIMMA(counts=counts, info=info, formula = " ~ treatment", block=kidney$replic)
}

