% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqHists.R
\name{pqHists}
\alias{pqHists}
\title{P- and q-value distributions plotted side-by-side}
\usage{
pqHists(x, what.p = "p.value", what.q = "q.value", alpha = 0.05,
  main = "significance distribution", ...)
}
\arguments{
\item{x}{matrix containing p-values and q-values}

\item{what.p}{the name of the column containing p-values}

\item{what.q}{the name of the column containing q-values}

\item{main}{The stem of the title, to which significance counts are attached}

\item{...}{additional arguments passed on to hist, like the color of the bars}
}
\description{
\code{pqHists} Make histograms of p-value and q-value distributions
}
\examples{
data(kidney)
counts<-kidney$counts
counts<-counts[sample(1:nrow(counts),1000),]
info<-data.frame(rep=kidney$replic, treatment=kidney$treatment)
stats<-pipeLIMMA(counts=counts, info=info, formula = " ~ treatment", block=NULL)
stats.fullmodel<-stats$stats
pqHists(stats.fullmodel, what.p="treatmenttumor_p.value",
 what.q="treatmenttumor_q.value", main="main effect treatment", breaks=100)

}

