% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPi0.R
\name{reportPi0}
\alias{reportPi0}
\title{Get pi0 from P-values}
\usage{
reportPi0(ps)
}
\arguments{
\item{ps}{a vector, data.frame or matrix of pvalues}
}
\value{
a vector (if is.numeric(ps)) or a matrix of pi0 and the number of affected elements
}
\description{
\code{reportPi0} Calculate pi0 from p-values using the qvalue package
}
\examples{
x <- sample(seq(0,1,by=0.0001),1000)
reportPi0(x)
df <- data.frame(x=sample(seq(0,1,by=0.0001),1000),
   y=sample(seq(0,.1,by=0.0001),1000),
   z=sample(seq(0,1,by=0.0001),1000))
}

