% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHeterosis.R
\name{testHeterosis}
\alias{testHeterosis}
\title{Bin genes by heterosis categories}
\usage{
testHeterosis(contrast.stats, alpha = 0.05, parentIDs = c("p1", "p2"))
}
\arguments{
\item{contrast.stats}{A dataframe containing the statistical output of contrasts between
parents and the F1. This must contain the following columns, in this order:
\itemize{
 \item{1. }{geneID: the gene identifier}
 \item{2. }{p1vf1.lfc: log2fold change of parent 1 vs. the f1}
 \item{3. }{p2vf1.lfc: log2fold change of parent 2 vs. the f1}
 \item{4. }{p1vp2.lfc: log2fold change of parent 1 vs. the parent 2}
 \item{5. }{p1vf1.fdrp: fdr-transformed p.value of parent 1 vs. the f1}
 \item{6. }{p2vf1.fdrp: fdr-transformed p.value of parent 2 vs. the f1}
 \item{7. }{p1vp2.fdrp: fdr-transformed p.value of parent 1 vs. the parent 2}
}}

\item{alpha}{FDR threshold for determining significance}

\item{parentIDs}{vector of length 2: the names of P1 and P2 respectively}

\item{...}{Not currently in use}
}
\value{
a dataframe where each row is one gene with the following columns:
\itemize{
 \item{1. }{geneID: the gene identifier}
 \item{2. }{highParent: p1 or p2}
 \item{3. }{order: the ranked order of p1, f1 and p2, separated by "."}
 \item{4. }{category1: the category of each gene following Paschold 2012}
 \item{4. }{category2: simplified category1 described in parentheses above}
}
}
\description{
\code{testHeterosis} Following categorization of Paschold 2012 (Genome Res.),
bin genes by their heterosis value. Tests are accomplished by contrasts (or similar)
}
\details{
Categorization is as follows:
##' \itemize{
 \item{1. }{LP<H<HP, no heterosis (additive)}
 \item{2. }{LP=H<HP, low parent (lp)}
 \item{3. }{LP<H=HP, high parent(hp)}
 \item{4. }{LP=HP=H, no significance (no DE)}
 \item{5. }{LP<HP<H, above high parent, parents are diff expressed (ahb)}
 \item{6. }{LP=HP<H, above high parent, parents are equal (ahb)}
 \item{7. }{H<LP<HP, below low parent, parents are diff expressed (blp)}
 \item{8. }{H<LP=HP, below low parent, parents are equal (blp)}
 \item{9. }{Abiguous (ambig)}
}
}
\examples{
\dontrun{
test<-testHeterosis(contrast.stats=contrast.stats,
  alpha = 0.1,
  parentIDs = c("FIL2","HAL2"))
}
}

