% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGeneRxn.R
\name{topGeneRxn}
\alias{topGeneRxn}
\title{Plots of expression vs. phenotype data.}
\usage{
topGeneRxn(v, info, sig, xdat = NULL, coldat = NULL, alpha = 0.05,
  nsig = 20, geneIDs = NULL, paletteChoice = NULL,
  pointcols = c("darkred", "forestgreen", "cornflowerblue"),
  ylab = "voom normalized expression", xlab = NULL,
  title = "comparison of expression and phenotype", ...)
}
\arguments{
\item{v}{Voom-transformed (or similar) expression data}

\item{info}{Experimental design dataframe or matrix}

\item{sig}{A vector of transformed (or not) p-values used to define
significance}

\item{xdat}{The column in the info dataset to plot on the x axis}

\item{coldat}{The column in the info dataset to color by, should be a factor}

\item{alpha}{A threshold for significance, currently not used.}

\item{nsig}{The number of genes to plot}

\item{paletteChoice}{If using a palette instead of predefined colors, which r color
brewer palette to use?}

\item{pointcols}{the colors for the factor (coldat). Needs to match exactly the number
of levels in coldat}

\item{ylab}{The label for the y-axis}

\item{xlab}{The label for the x-axis}

\item{title}{The title for the plot}

\item{...}{additional arguments passed on to geom_point()}
}
\value{
a table with the data used to make the plot
}
\description{
\code{volcanoPair} Make a matrix of plots for a set of genes. Typically,
for a given effect, pull out the strongest genes and plot their phenotype vs.
voom normalized gene expression data. Current implementation works best for
continuous x data.
}
\details{
Takes a gene expression matrix, subsets by the nsig most significant genes,
melts to the long format, then plots via ggplot in a scatterplot matrix.
}
\examples{
data(kidney)
counts<-kidney$counts
counts<-counts[sample(1:nrow(counts),1000),]
info<-data.frame(rep=kidney$replic, treatment=kidney$treatment)
stats<-pipeLIMMA(counts=counts, info=info, formula = " ~ treatment + rep", block=NULL, getTopTable=T, getEbayes=T)
stats<-stats$stats

}

