% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPair.R
\name{volcanoPair}
\alias{volcanoPair}
\title{Scatter plot of log2 fold changes.}
\usage{
volcanoPair(lfc1, lfc2, sig1, sig2, alpha = 0.05, pointcols = c("darkred",
  "skyblue", "forestgreen", rgb(0, 0, 0, 0.5)), pchs = c(19, 19, 19, 19),
  cex = c(0.5, 0.5, 0.5, 0.5), legpos = "topleft", ...)
}
\arguments{
\item{lfc1}{A vector of log-2 fold change (or similar data) for the x axis}

\item{lfc2}{A vector of log-2 fold change (or similar data) for the y axis}

\item{sig1}{Either a 0/1 binary vector of significance, or a
vector of transformed p-values for coloring the x axis points}

\item{sig2}{Either a 0/1 binary vector of significance, or a
vector of transformed p-values for coloring the y axis points}

\item{alpha}{If providing a vector of transformed p-values,
this specifies the threshold for siginficance. Otherwise, not used.}

\item{pointcols}{Vector of legnth 4. Indicates  the color of significant points
for both [1], just the y axis [2], just the x axis [3] or neighter.}

\item{pchs}{Follows pointcols for point shapes}

\item{cex}{Follows pointcols for point sizes}

\item{legpos}{The position of the legend. Defaults to the top right.}

\item{...}{additional arguments passed on to plot}
}
\value{
a table with the number of significant effects
}
\description{
\code{volcanoPair} Make an expression volcano plot from log2-fold change and p-value data
}
\details{
Using base R graphics, this function tabulates the number of significant values and plots
}
\examples{
\dontrun{
data(kidney) #from the simseq package
counts<-kidney$counts
counts<-counts[sample(1:nrow(counts),1000),]
info<-data.frame(rep=kidney$replic, treatment=kidney$treatment)
stats<-pipeLIMMA(counts=counts, info=info, formula = " ~ treatment + rep", block=NULL, getTopTable=T, getEbayes=T)
stats<-stats$stats
volcanoPair(lfc1=stats$treatmentTumor_logFC,
           lfc2=stats$rep6090_logFC,
            sig1=stats$ebayesQvalue_treatmentTumor,
            sig2=stats$ebayesQvalue_rep6090)
}

}

