% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voom2PCA.R
\name{voom2PCA}
\alias{voom2PCA}
\title{PCA of voom-transformed counts.}
\usage{
voom2PCA(v, info, ids, plotit = TRUE, pcas2return = 3,
  plot.cols = "black", ...)
}
\arguments{
\item{v}{Counts matrix, typically transformed by limma::voom.
Possibly output from pipelimma, in the slot "voom".}

\item{info}{The experimental design information matrix}

\item{ids}{A vector of the individual names}

\item{plotit}{Logical, should the pca be plotted?}

\item{...}{additional arguments passed on to pairs, for example, the colors of bars}
}
\value{
a list containing a dataframe with the experimental design data, merged with the 1st
3 principal component axes and a vector of %variance explained by PCA axes.
}
\description{
\code{voom2PCA} Run principal component analysis on matrix of voom-normalized counts
}
\details{
This function uses the R function princomp to calculate principal components
}
\examples{
data(kidney)
counts<-kidney$counts
counts<-counts[sample(1:nrow(counts),1000),]
info<-data.frame(rep=kidney$replic, treatment=kidney$treatment)
### Not Run ### stats<-pipeLIMMA(counts=counts, info=info, formula = " ~ treatment", block=NULL)
### Not Run ### pc <- voom2PCA(v=stats$voom[["E"]], info=info, ids=rownames(info),plotit=TRUE)
### Not Run ### library(ggplot2)
### Not Run ### ggplot(pc, aes(x=PC1, y=PC2, col=treatment))+geom_point()
}

