% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeTopGo.R
\name{pipeTopGo}
\alias{pipeTopGo}
\title{Pipeline for GO enrichment analysis}
\usage{
pipeTopGo(geneID2GO, genes.of.interest, nodes4table = NULL,
  nodes4graph = NULL)
}
\arguments{
\item{geneID2GO}{A go annotation (list). Each element is the name of a gene,
containing a vector of GO annotations for that gene}

\item{genes.of.interest}{A vector of genes to look for GO enrichment}

\item{nodes4table}{The number of GO terms to print in the top table.
Defaults to the number of GO terms with fisher test p.value <= 0.05.}

\item{nodes4graph}{The number of GO terms to display in the plot
#' Defaults to the number of GO terms with fisher testp.value <=0.05.}
}
\value{
a table with the significant (or top n) GO terms and a directed network
of GO terms
}
\description{
\code{pipeTopGo} Invokes the R package topGO to run gene ontology enrichment analysis
Takes a vector of gene names and a topGO custom (or not) enrichment annotation.
}
\examples{
data(geneID2GO)
geneNames <- names(geneID2GO)
myInterestingGenes <- sample(geneNames, 100)
test<-pipeTopGo(geneID2GO=geneID2GO, genes.of.interest=myInterestingGenes)
}

